// packages/js/nlux/src/types/markdown/markdownElement.ts
export type MarkdownElementName = 'Root'
    | 'Paragraph'
    | 'LineBreak'
    | 'BoldAsterisk'
    | 'ItalicAsterisk'
    | 'BoldUnderscore'
    | 'ItalicUnderscore'
    | 'Blockquote'
    | 'Heading1'
    | 'Heading2'
    | 'Heading3'
    | 'Heading4'
    | 'Heading5'
    | 'Heading6'
    | 'OrderedList'
    | 'UnorderedList'
    | 'Code'
    | 'CodeBlock'
    | 'HorizontalRule'
    | 'Link'
    | 'Image';